SDATqap1/src/Main/Java/com/example/Library.java
package com.example;

import java.util.ArrayList;
import java.util.List;

public class Library {
    private List<Book> books;
    private List<LibraryMember> members;

    public Library() {
        books = new ArrayList<>();
        members = new ArrayList<>();
    }

    public void addBook(Book book) {
        books.add(book);
    }

    public void removeBook(Book book) {
        books.remove(book);
    }

    public List<Book> getBooks() {
        return books;
    }

    public void registerMember(LibraryMember member) {
        members.add(member);
    }

    public List<LibraryMember> getMembers() {
        return members;
    }

    public boolean isBookAvailable(Book book) {
        return books.contains(book);
    }

    public Book findBook(String title) {
        for (Book book : books) {
            if (book.getTitle().equalsIgnoreCase(title)) {
                return book;
            }
        }
        return null;
    }
}
